+++ FantasmoDemoApp/app/build.gradle
@@ -75,10 +75,10 @@ dependencies {
     implementation 'androidx.legacy:legacy-support-v4:1.0.0'
 
     // Fragment Navigation
-    implementation "androidx.navigation:navigation-fragment-ktx:2.3.0"
+    implementation "androidx.navigation:navigation-fragment-ktx:2.3.5"
 
     // Google ARCore
-    implementation 'com.google.ar:core:1.24.0'
+    implementation 'com.google.ar:core:1.27.0'
     implementation 'com.google.ar.sceneform.ux:sceneform-ux:1.17.1'
     implementation 'com.google.ar.sceneform:core:1.17.1'
     implementation 'com.google.ar.sceneform:animation:1.17.1'
