+++ FantasmoSDK/app/src/main/res/layout/fmparkingview.xml
@@ -64,6 +64,7 @@
                     android:text="@string/anchor"
                     android:textColor="#000000"
                     android:textSize="18sp"
+                    app:theme="@style/ColorAnchorSwitchStyle"
                     app:layout_constraintBottom_toTopOf="@+id/checkParkingButton"
                     app:layout_constraintStart_toStartOf="@+id/localizeToggle"
                     tools:ignore="UseSwitchCompatOrMaterialXml" />
@@ -78,6 +79,7 @@
                     android:text="@string/localize"
                     android:textColor="#000000"
                     android:textSize="18sp"
+                    app:theme="@style/ColorLocalizeSwitchStyle"
                     app:layout_constraintBottom_toTopOf="@+id/anchorToggle"
                     app:layout_constraintStart_toStartOf="parent"
                     tools:ignore="UseSwitchCompatOrMaterialXml" />
