+++ FantasmoSDK/app/src/main/res/values/styles.xml
@@ -1,0 +1,27 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <style name="ColorLocalizeSwitchStyle">anchorSwitch
+        <!-- colorPrimary is used for the default action bar background -->
+        <!--item name="colorPrimary">#4fff4f</item-->
+
+        <!-- colorPrimaryDark is used for the status bar -->
+        <item name="colorPrimaryDark">#f1f1f1</item>
+
+        <!-- colorAccent is used as the default value for colorControlActivated,
+                 which is used to tint widgets -->
+        <item name="colorAccent">#ea4335</item>
+
+        <!-- You can also set colorControlNormal, colorControlActivated
+        colorControlHighlight, and colorSwitchThumbNormal. -->
+    </style>
+
+    <style name="ColorAnchorSwitchStyle">anchorSwitch
+        <!-- colorPrimaryDark is used for the status bar -->
+        <item name="colorPrimaryDark">#f1f1f1</item>
+
+        <!-- colorAccent is used as the default value for colorControlActivated,
+                 which is used to tint widgets -->
+        <item name="colorAccent">#00FFFF</item>
+    </style>
+
+</resources>
